requirements.txt
amqp==2.6.1
apns2==0.7.2
asgiref==3.3.4
automium==0.2.6
automium-web==0.1.1
Babel==2.8.0
backcall==0.2.0
billiard==3.6.4.0
boto3==1.12.6
botocore==1.15.49
celery==4.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.7
Django==3.1.7
django-filter==2.4.0
django-safedelete==0.5.5
django-storages==1.9.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-yasg==1.20.0
Flask==1.1.2
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
importlib-metadata==4.0.1
inflection==0.5.1
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==4.6.11
legacy==0.1.6
MarkupSafe==1.1.1
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
prompt-toolkit==3.0.18
psycopg2==2.8.4
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==1.6.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.23.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.3.7
sentry-sdk==0.17.0
six==1.15.0
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.11
views==0.3
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.1
